remove implementation code for origin map
The "origin map" was the very first way Sourcegraph supported repo syncing. It was a list of mappings of the form "a Sourcegraph repository named ghe.example.com/$REPO should be cloned from https://ghe.example.com/$REPO.git". It was deprecated pre-2.8 and removed in 2.9 from site config. The impl code still existed for it (which is what this commit removes).

- Removes ORIGIN_MAP env var support (it was still possible to set this env var and use the origin mapping functionality, even though it was removed from site config)
- Removes the impl code for origin mapping