search: Create a copy of zoekt's query pkg
We will use it as a base for our new parser and query type. It is superior to
our Pattern struct since it encodes a general expression tree rather than a
flat list of options.

Note: This is a copy from our fork of zoekt, so already includes some
Sourcegraph specific changes.