accept access tokens in UI routes, not just in API routes
Prior to this change, only `/.api` routes would allow authentication using an
access token. This was purely because those were the only routes that actually
needed access token authentication.

However, with the raw file API I am adding at e.g. `/github.com/gorilla/mux/-/raw/mux.go`
it is desirable to allow access token authorization. This change does just that.

This cli package is not tested, so adding tests here is non-trivial and would
require a lot of refactoring or laying a lot of groundwork (I suspect 1d of
work). I've opted to not do this.