search: Add ExpandRepo (#868)
Our indexed search backend uses RepoSets (an exact set of repositories) to do
searching. ExpandRepo allows us to take our Repo queries and turn them into
sets for the backend. This also ensures that we only return results for
repositories that Repos.List returns, which in turn ensures we respect ACLs.

In the future we will ensure the Repo token on the indexed search backend uses
the same logic as our Repo token. Then we can remove this function and instead
send the query `(and RepoSetFromReposDotList q)`.