remove siteID from site config
The siteID value is auto-generated upon initialization. There is little value in being able to configure it manually. There is no guarantee that any site uses the same value, and there is no enforcement (eg remote enforcement) of it and there never will be.

For instances that set siteID previously, they may not have a `site_config` row. This commit also changes the behavior when that occurs. Next time the server restarts, a `site_config` row will be created automatically, and the site will be marked as initialized if any users exist. This means there will be no interruption to users. (Previously, in this case, users would see a screen telling them to ask their site admin to initialize the server.)

The env var TRACKING_APP_ID is still respected, for backcompat with the legacy way this is set on Sourcegraph.com and dogfood.