rename appURL to externalURL (#924)
The name "external URL" is more descriptive than "app URL". (What is an app? Which app? Is this the internal application server?) It is also consistent with the overall style of our site config properties, where full names are used over abbreviations.

The env var `SRC_APP_URL` still sets this value. That legacy way of providing site config will soon go away, so it did not seem worth changing in this commit.


> This PR updates the CHANGELOG.md file to describe any user-facing changes.