SECURITY: allow UI endpoints to use access token authentication (take 2) (#951)
* accept access tokens in UI routes, not just in API routes

Prior to this change, only `/.api` routes would allow authentication using an
access token. This was purely because those were the only routes that actually
needed access token authentication.

However, with the raw file API I am adding at e.g. `/github.com/gorilla/mux/-/raw/mux.go`
it is desirable to allow access token authorization. This change does just that.

This cli package is not tested, so adding tests here is non-trivial and would
require a lot of refactoring or laying a lot of groundwork (I suspect 1d of
work). I've opted to not do this.

* allow specifying access tokens via basic auth

Prior to this commit, it was possible to specify access tokens via a query
parameter:

	https://sourcegraph.com/foobar?token=abcdef

As well as via the `Authorization` header:

	Authorization: token abcdef

After this change, it is now possible to also specify the token via basic
authentication, i.e.:

	https://abcdef@sourcegraph.com/foobar

This change is desired for cosmetic reasons, but is backed by the same security
guarantees and unit tests that already exist for the query parameter method.