search: Find revisions in hierarchical query
We have a seperate token to specify refs to search. This allows a more
powerful model where we can AND and OR the branches to search. It also leads
to a less confusing syntax around the repo names. This commit allows us to
inspect a query to find out which references we need to search.

Note: Currently the token is "branch", but this will likely change. As such
the language used in this commit is ref since that is the correct language to
use for resolving git ref specs.