support marking extensions as work-in-progress
Problem: it's not clear which extensions on Sourcegraph.com are ready to be used. There is a convention to add "WIP" to the title, but that is not clearly communicated and still results in the extension being shown to users prominently.

An extension with no published releases or whose title begins with `WIP:` or `[WIP]` is considered a work-in-progress extension. WIP extensions:

- are not shown in the initial list of extensions on the extension registry (they are only shown when the user has typed in a query)
- are sorted last on the extension registry list of extensions when there is a query (unless the user has already added the WIP extension)
- have a red "Work in progress" badge on the extension card in the list
- have a red "Work in progress" badge on the extension's page

fix #480