remove panel title
The panel title showed the first hover message. This is not a well-defined concept and sometimes resulted in the panel title being very tall, which completely messed up the layout of the page. Also, there were bugs in the implementation, such as the loading indicator continuing to show up sometimes.

fix #915