remove panel title (#1011)
The panel title showed the first hover message. This is not a well-defined concept and sometimes resulted in the panel title being very tall, which completely messed up the layout of the page. Also, there were bugs in the implementation, such as the loading indicator continuing to show up sometimes.

fix #915 

The panel title is the thing that has `var debug bool` above `INFO   DEFINITION`:

![screenshot from 2018-11-15 00-55-27](https://user-images.githubusercontent.com/1976/48541296-300e6b80-e871-11e8-9498-9d6e2cac9d07.png)

It is now removed.

No panel title anymore:

![screenshot from 2018-11-15 00-54-01](https://user-images.githubusercontent.com/1976/48541204-fd647300-e870-11e8-87c3-28c00121906e.png)