add RegistryExtension.publishedAt date, bump updated date after publishing release (#1007)
Previously, an extension's "last updated" date on its registry extension page showed the last time its metadata was updated. The only thing that would bump this date up is creating or renaming the extension. This date was misleading because people understandably expected the "last updated" date to be bumped up when a new release of the extension was published. This commit implements that behavior.

fixes #623