feat: expose workspace roots in extension API (#991)
* feat: expose workspace roots and file system API

These new APIs make it easier for language extensions to communicate with language servers, which:

- generally have a 1-to-1 correspondence with workspace roots
- need the file contents in textDocument/didOpen

The `sourcegraph.FileSystem` API is intentionally very limited (i.e., it lacks a way to read a directory). It will be expanded in the future as needed.

* update docs

* update workspace roots in extensions environment in browser extension

This makes the "workspace root" (the repo being viewed on the code host, plus its revision) available to the extension. Previously the extension only had access to the `visibleTextDocuments`.