remove unneeded deps that were only used for LSP types (#1033)
* remove old LSP server capabilities code

* remove vscode-languageserver dep

Use rxjs Unsubscribable instead of vscode Disposable. This, and the other changes in this commit, lets us remove the (indirect) dependency on vscode-languageserver from our package.json (in a subsequent commit).

* remove javascript-typescript-langserver dependency

Copies over a few types from the package that were still in use. These types will be removed in Sourcegraph 3.0 (due to the migration to extension-based cross-repository references).


* remove vscode-languageserver-types dep

* remove dep on vscode-jsonrpc

This package was no longer around anyway.