db: rename "site config" to "global state"
This database table is used for storing a single row with values indicating the
generated site ID and whether or not the site is initialized.

Both values in this table currently are what I would consider to be 'global
state pertaining to Sourcegraph', and as such I believe it should be named
'global state' instead. This also allows makes it more obvious that other
global Sourcegraph state could one day go into this table.

The primary motivation for making this change is because "site config" is a
name which is very easy to confuse with the "site config JSON" and there is
in fact zero relationship between the two (and there can't be even when we
move site config to the DB, because that requires multiple rows and this global
state does not).