mv cmd/frontend/db/dbconn to pkg/dbconn (#1084)
* mv dbconn to pkg/dbconn

With the introduction of the management console, the frontend will no longer be
the only service which talks to the database. Therefor, sharing the logic to
connect to and migrate the database makes sense. This change does just that.

It was also sharable before, but it would have meant importing
`cmd/frontend/db/dbconn` from the management console which does not make sense
logically as they are separate services.

Migration logic still lives in pkg/dbconn, because we only have one database
(not multiple).

Helps #965

* mv cmd/frontend/db/migrations -> migrations/

Previously the generated contents of the migrations/ folder was stored
in a subpackage of the frontend. This no longer makes sense with pkg/dbconn
importing it.

Additionally, this is something that confuses me every time I add a migration; i.e.
"what directory do I generate and why is it not the same as the directory I modified?!"

* dev: use go generate -x flag for improved CI debugging

* pkg/dbconn: add disclaimer about most services using api.InternalClient instead

* dev/check: add check to ensure pkg/dbconn is only imported by allowed services

* update some imports + editor autofmt query (removes mixed spaces / tabs, uses just tabs now)

* NOCHANGELOG