search: Support hierarchical queries in searcher
We add an evaluator for hierarchical queries in searcher. It is based off how
Zoekt does evaluation (indexed search). We construct a tree which mirrors the
query, but statefully collects matches for a document. This matchtree is then
run for each document.

Our searcher code skips running a regexp on a document if it doesn't contain a
substring from the regular expression. This is achieved in a matchtree by
creating a subtree matcher which looks for the string, and anding it with the
regex searcher. This is the reason for "cost" as part of matching and
"novisit" to avoid collecting those matches.

This commit is missing exposing the RPC interface, as well as performance
work.