xlang: Remove AsyncHandler from TestCache
AsyncHandler introduces flakiness since each request will go into a goroutine.
Given the test is doing sets followed by gets we need the server to run in the
same order. With AsyncHandler it will be upto the scheduler when a request's
goroutine runs. This also means we can remove the 50ms sleep.

Fixes #797