support showing panel on all pages
Previously, the panel would only be displayed on the blob or tree pages. Now it can appear on any page. There are no panel views yet that are intended to show up on other pages, but it is now possible.

This change also makes it so that the panel takes up the full screen width, even on tree/blob pages. Previously it only took up the partial width starting at the sidebar (if open) on the left. Having more width in the panel is more important than having more height on the sidebar, so this change is beneficial. (It also is much simpler to implement this way rather than continuing to have the panel take partial width on tree/blob page, because it allows the panel to be rendered in a single place high up in the React component hierarchy.)