pkg/endpoint: Simpler implementation of GetAll
I over-engineered the crap out of my last commit. I was worried about the case
of a very large amount of keys so implemented a smart way to get the
values. However, the algorithm was O(nlogn + nlogk). For large n this would
just be O(nlogn). However, the naive solution is O(nlogk) :/. So I just
implemented something slower in the degenerate case.

Instead we provide a helper which calls get N times, but does it with only
acquiring the lock once.