search: Expand ref to repo@commit for searcher
Previously I designed the API such that we would pass in repo@commit as search
options. However, that means we block search anything until all of those
commits are resolved. So the API will be adjusted such that we only pass in a
list of repository names (possibly even types.Repo). This is the first and
most intense change. It involves partially evaluating a query to understand
which ref patterns apply to which repositories.

I imagine other service will want to know answers to these sort of questions
when we have search extensions. When that time comes, we can factor this sort
of logic out into an exported API.