Don't close dbconn.Global after running migrations
This commit changes the previous `dbconn.DoMigrateAndClose` function to
`dbconn.DoMigrate`, which doesn't close the underlying global database
connection.

This became a problem after migrating github.com/golang-migrate/migrate to v4.1.0.
Specifically, this commit https://github.com/golang-migrate/migrate/commit/1e4c50b70d6a45a46ca4b7f1a29f4e62c52c2cde,
which changed the `Postgres` driver `.Close()` method to also close the
underlying connection pool `*sql.DB`, instead of the single `*sql.Conn`
as before.

Since we have a shared global `dbconn.Global` connection pool, and we expect it to
remain open for the life-time of the process, we should not be closing
it after performing migrations.

Here's an example of a place that expected `dbconn.Global` to be open, but failed
with `Table schema_migrations not found: sql: database is closed` after the dependency upgrade.
It's a function that is meant to be ran between tests that use the Postgres database.

    // pkg/db/dbtesting/dbtesting.go
    func emptyDBPreserveSchema(d *sql.DB) error {
        _, err := d.Exec(`SELECT * FROM schema_migrations`)
        if err != nil {
                return fmt.Errorf("Table schema_migrations not found: %v", err)
        }
        return truncateDB(d)
    }

NOCHANGELOG