remove unused code from browser extension
This removes a lot of old pre-newInject code. None of this was in use anymore. The purpose of removing it is to make it easier to make extensions work well with the browser extension (by limiting the number of places that need to invoke the extensions client API).

Process:

1. Remove `export` keyword from definitions that were only used in the same file
2. Remove definitions that had no usages
3. Repeat

NOCHANGELOG