graphqlbackend: New search feature flagged
The new implementation requires much less code, most of which can be factored
into another package. The idea being to keep as much logic out of
graphqlbackend as possible.

We gate the feature behind a simple `!hier!` prefix in search queries. This
will allow us to test in dogfood and production on real datasets to find
potential issues.