cmd/management-console: make asset generation implicitly build assets first
Prior to this change it was required to run pre-build.sh before running
`go generate` otherwise you would end up with assets missing everything.

After this change, `go generate` is the authoritative source and it builds
the assets first.

This fixes a bug with the latest 3.0.0-alpha image where it would not contain
the management console assets, thus making it impossible to access.