remove lodash-es, just use lodash
lodash-es causes problems with jest (see https://github.com/facebook/create-react-app/issues/5241#issuecomment-426269242) and it is unnecessary for us to use it. We historically used it because it could be tree-shaken better than lodash, but now that we are using the babel lodash plugin, we get the same tree-shaking when we use plain lodash.