support typed property values in Context (#1459)
This allows the context (which is used for `when` expressions and other expressions in an extension manifest) to contain properties whose value is a TypeScript type other than the primitive types (or Context itself). This is useful when, for example, you have a `TextDocumentPositionParams` value that is the value of a context property; using `Context<TextDocumentPositionParams>` makes it typecheck.

There is no user-facing behavior change. This is just a change to types, not impl code.

Part of #1313 but useful by itself.