store non-resolved input revision for workspace roots
This makes it so that the original input revision (e.g., a branch name) for each workspace root is stored. Previously only the full Git commit SHA was available (in the workspace root URI).

This is a prerequisite to avoid navigating the user from a URL with a branch name to a URL with a full Git commit SHA when they perform actions such as go-to-definition.