sort dirs/files top-bottom in columns on tree page
Previously, tree entries (files and dirs) were sorted left-right, which was hard to read. Now they are sorted top-bottom, which is more familiar and easier to read.

This also removes the usage of CSS grid layout favor of CSS column layout. There are some edge cases, such as column sparseness on wide screens with a small number of entries. To completely fix this, it is necessary to use JavaScript (which is not worth it right now due to the added complexity).