use React error boundaries to avoid blank pages upon unexpected errors (#1586)
If a React component throws an error during rendering, previously the entire page would be blank. This commit introduces [React error boundaries](https://reactjs.org/docs/error-boundaries.html) so that the error is confined and a nice(r) error page is displayed.

This is NOT intended to handle errors that can be anticipated. Components must do their own error handling. It is just a better worst-case when an error is not handled correctly.

fix #659

![screenshot from 2018-12-30 06-45-07](https://user-images.githubusercontent.com/1976/50547486-e9e25100-0bff-11e9-9e01-85cdbf70803f.png)
![screenshot from 2018-12-30 06-43-16](https://user-images.githubusercontent.com/1976/50547487-ec44ab00-0bff-11e9-85b8-f7903ad3a1fe.png)
![screenshot from 2018-12-30 06-41-51](https://user-images.githubusercontent.com/1976/50547489-eea70500-0bff-11e9-9e56-2831bf5471a8.png)
![screenshot from 2018-12-30 06-41-38](https://user-images.githubusercontent.com/1976/50547490-efd83200-0bff-11e9-934a-1c1cbec8b29e.png)