doc/admin: add simpler instructions for resetting management console password
If you still have access to Sourcegraph, there is an easier way to generate a
new password that does not require you bcrypting it. Document this as the first
option but keep the more manual method as a fallback in the event that your
Sourcegraph instance is broken for some reason.