show "Reload required" error if app JS is redeployed after user's initial page load (#1803)
* use contenthash not chunkhash

Webpack recommends using contenthash because it is based on the hash of the actual chunk content, not of the chunk source, so it is more accurate and changes only when a meaningful change has occurred. There is no difference in actual outcomes for our use case currently.

* reset error page when user navigates

This lets the user navigate away from the error. For example, if an error occurs on one tab in the user area, they can click on other tabs and those will work. Previously, the error would remain even when the user attempted to navigate away from it, as long as the error's parent stayed the same.

* show "Reload required" error if app JS is redeployed after user's initial page load

fix #1760