support category:, tag:, and #-keywords for queries to remote extension registry (#1832)
The query to the remote extension registry was being treated as a literal query string and was not processed for category:, tag:, and #-keywords. This meant that self-hosted instances using those tokens in the extension registry query would not get the right results from the remote extension registry.

fix #1830

There needs to be better integration testing of fetching and merging data from the local and remote extension registries (https://github.com/sourcegraph/sourcegraph/issues/1831).