suppress log noise from config fetching when frontend is temp unavailable (#1828)
When starting up Sourcegraph (sourcegraph/server, the cluster, and the dev server), there is a lot of logspam before it shows the "Sourcegraph is ready at" message. This will make users think that there is an error, when in reality it's *usually* just that the other services are waiting for the frontend to be ready (which can take a few seconds when it needs to migrate the DB).

A slightly better fix might be for the frontend to immediately start listening for conf requests when it starts up (even before DB migration is done). Then the delayBeforeUnreachableLog could be lowered from its current default of 15s, and actual errors would be printed sooner. That is more complex and is not necessary right now.

fix #1826