Refactor IsAllowedURL into Client struct
Refactor the IsAllowedURL to be a method on the new Client struct.

This refactoring enables using the Client with different configurations,
and since the Client must understand how to talk to StackExchange
compatible APIs it makes sense for the same client which speaks to the
APIs to express whether or not it can support those APIs.

Client allowList and the default list of allowed URL patterns are not
exported past the package boundary (yet?).