Refactor Client configuration to functional options
This makes it easier to construct non-standard Clients (e.g for testing
with absurdly short timeouts) without struggling against a broad API.

This commit also introduces a LockMechanismer which will allow the
Client to be configured with various ways to manage locks to prevent
concurrent accessing of URLs.

Some previously unexported fields are now exported because they now
have a value to anyone who may be assembling a custom Client for their
use-case.