Introduce a simple in-process sync.Locker
This commit introduces code to guard against concurrent requests (within
a single Client) to the same URL resource.

The locking mechanism is pluggable, and the tests and implementation
here demonstrate two models. A completely fake pre-locked locker, which
is used to test the behaviour of Client.FetchUpdate in cases of lock
contention, and a real implementation which uses a simple sync.Mutex.

Further implementations could easily leverage remote locks such as
atomic keys in Redis, or similar.

As sync.Locker.Lock() blocks indefinitely the lock wait is handled in a
goroutine which will/must eventually yield. Successfully attaining a
lock will cause a context.Context to be cancelled (signalling cancelling
the waiting to get a lock) and this propagates to a select{} block which
checks whether the goroutine waiting for a lock reports
context.DeadlineExceeded or not.