Refactor to begin to make HTTP requests.
Next steps from here are to make sure that the URL given is mutated into
a StackExchange API URL and has the parameters injected into it, and
finally to make the request and parse the response.

I'm less than happy with the httpDoFn, I would have preferred to wrap a
whole client, or find a less goofy name, but it seems to be the smallest
API I need, so far and makes the tests significantly easier to read.