server: Use include directives for nginx to create an upgrade path (#1907)
We currently generate the nginx.conf for users, but then it is controlled by
users. If we want to change a setting in nginx.conf (such as changing a proxy
timeout), we then have to rely on user instructions to update the config.

Instead we can hook into the config via well placed include directives.