all: Replace godockerize with Dockerfiles
We use a branch of godockerize which generates a build.sh and Dockerfile instead
of directly building the docker image.
https://github.com/sourcegraph/godockerize/tree/gen-build

This was run on all main entrypoints (excluding a few which already have a
build.sh / shouldn't be docker images):

  go list -f '{{ if (eq .Name "main") }}{{.ImportPath}}{{ end }}' ./... \
    | grep cmd | grep -v goreman | grep -v 'schemadoc' | grep -v /server \
    | while read pkg; do godockerize build $pkg; done

The build.sh should likely be refactored into using a shared shell script. But
for now this gives us a nice and easy to change replacement.

This commit also removes godockerize from CI and our go.mod.