frontend: Allow setting CACHE_DIR
In production we are using up the scratch space for tmp due to archives and the
raw endpoint. This allows us to change the location to store the archives, such
as we do in other services that use archives. This will be used to configure a
path to a local SSD in production.