cmd/frontend/internal/cli: make bad externalURL configuration abide by guidelines
Bad critical configuration should never take down the server, as doing so would
lock the admin out of the management console. Instead, the server must just
'do nothing' until it is fixed and the server restarted. After this change, we
properly abide by this rule for invalid `externalURL` configuration (the last
instance of config code which did not obey this rule).

There is a remaining open issue for me to properly document this rule and
other Sourcegraph config development guidelines.

Fixes #1431