Syntax highlighting for Kotlin+Dart; various syntect_server maintenance  (#2270)
* cleanup syntect launch scripts

All of these env vars are already set in the Docker image itself for a long
time now, we do not need to specify them here. Also, we do not need a separate
launch script for this.

* cmd/server: update syntect_server to include Kotlin+Dart support (and remove unused update script)

* dev: pin the syntect_server version

Using latest always is not good, as there could theoretically be backwards
incompatible changes in syntect_server which would break dev environments.