remove empty "Experimental configuration" section (#2293)
The browser ext options menu "Experimental configuration" section was empty. The configuration section only had "Allow error reporting". Instead of refactoring this code to selectively show the experimental section when it would be non-empty, this commit just removes it altogether. If a new experimental configuration option is added, it is OK for it to be in the "Configuration" section (and it can be annotated with "EXPERIMENTAL" if needed).