remove hover priority from extension API (#2039)
The hover priority was previously used to prevent showing *both* the basic-code-intel hover and the precise hover (for TS/JS/Go/Python). #2035 makes this unnecessary, because each language extension will be the sole source of hovers for that language (and they will use the basic-code-intel extension code as a library).

~~This is blocked on #2035; do not merge this until #2035 is merged.~~ Merged, unblocked

Backcompat: Extensions that set the `Hover#priority` property will continue to work. That value will be ignored, and it being set will not cause any serious issues. Any instances on 3.0 beta using basic-code-intel will see duplicate hovers, but that is acceptable for beta (we do not have a backcompat guarantee between 3.0 beta and 3.0 for extensions).

fix #1938

Partially reverts 3c077164ad31d70ed314183a50c72b5f04e681de.