Fix: code detection logic in searchResultMatch component
We relied on the markdown renderer to produce <code> tags to wrap codeblocks
when rendering diff and commit results. Our markdown rendering changed, so this
is no longer the case.

TODO: add tests for this so it doesn't break again.
TODO: don't depend on the same markdown library as our docs site