simplify "Is Sourcegraph.com?" check
I wanted to write a test for Sourcegraph.com and non-Sourcegraph.com behavior for something, so I wanted to mock the result of SourcegraphDotComMode. It is simpler to mock it if we remove the special-case checking of the URL. That special-case is not necessary anyway because Sourcegraph.com is deployed with SOURCEGRAPHDOTCOM_MODE=true (I confirmed this with `kubectl -n prod describe deployment/sourcegraph-frontend` on the Sourcegraph.com k8s cluster).

Also standardizes other places where we mock this and provides a documented/standard way to mock it.