migrations: New migrations use transactions
We discovered that the migration library does not use transactions, nor has it
ever used transactions. The only place it uses transactions is for SetVersion,
not the actual migrations. The [README][1] for the postgres driver encourages
the use of transactions for multi-statement migrations. We believe it is better
to just always use transactions.

We added a template for the created migrations. It uses a stronger transaction
type (ISOLATION LEVEL SERIALIZABLE) and it includes comments with advice from
the README.

[1]: https://sourcegraph.com/github.com/golang-migrate/migrate@v4.2.5/-/blob/database/postgres/README.md?view=code#L26