Minimize Prop drilling
Prop drilling refers to having to pass-through the same props through a
lot of components. This removes duplicated Prop declarations by sharing
common Prop interfaces and uses the {...props} spread syntax to pass
through bags of props. This eliminates the need to add each new prop
that needs to be passed through to a lot of components on every change.